requirements.txt
typing-extensions
pillow
aiohttp
bs4
matplotlib
numpy
